app_api/app_images/cloud_run_container_app_images/v1/pyproject.toml
[tool.poetry]
name = "app-images"
version = "0.5.0"
description = ""
authors = ["Alex Girlea"]

[tool.poetry.dependencies]
python = "~3.10"
uvicorn = "~0.22"
fastapi = "~0.99"
cryptography = "~42.0"
python-multipart = "~0.0.16"
python-magic = "~0.4.27"
pillow = "~10.3"
cairosvg = "~2.7"
opencv-python =  "~4.10.0.82"
pytesseract = "~0.3.10"
pyzbar = "~0.1.9"
svgwrite = "~1.4.3"
python-barcode = "~0.15.1"
rembg = "~2.0.57"
SimpleITK = "~2.4.0"
moviepy = "~1.0.3"
numpy = "~1.26.4"
google-cloud-secret-manager = "~2.20.2"

# ml model
onnxruntime = "~1.20.1"

#logging
rich = "~13.9.4"

# this if for opentelemetry
opentelemetry-api = "~1.24.0"
opentelemetry-exporter-gcp-monitoring = "~1.6.0a0"
opentelemetry-exporter-gcp-trace = "~1.6.0"
opentelemetry-exporter-otlp = "~1.24.0"
opentelemetry-exporter-otlp-proto-common = "~1.24.0"
opentelemetry-exporter-otlp-proto-grpc = "~1.24.0"
opentelemetry-exporter-otlp-proto-http = "~1.24.0"
opentelemetry-instrumentation = "~0.45b0"
opentelemetry-instrumentation-asgi = "~0.45b0"
opentelemetry-instrumentation-fastapi = "~0.45b0"
opentelemetry-proto = "~1.24.0"
opentelemetry-sdk = "~1.24.0"
opentelemetry-semantic-conventions = "~0.45b0"
opentelemetry-util-http = "~0.45b0"

[tool.poetry.group.dev.dependencies]
pytest = "~8.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
